chore(test): extract unixfs test utils to go-unixfsnode
Ref: https://github.com/ipfs/go-unixfsnode/pull/47
