+++ pkg/internal/itest/http_fetch_test.go
@@ -19,10 +19,10 @@ import (
 	datatransfer "github.com/filecoin-project/go-data-transfer/v2"
 	"github.com/filecoin-project/lassie/pkg/internal/itest/mocknet"
 	"github.com/filecoin-project/lassie/pkg/internal/itest/testpeer"
-	"github.com/filecoin-project/lassie/pkg/internal/itest/unixfs"
 	"github.com/filecoin-project/lassie/pkg/lassie"
 	httpserver "github.com/filecoin-project/lassie/pkg/server/http"
 	"github.com/ipfs/go-cid"
+	unixfs "github.com/ipfs/go-unixfsnode/testutil"
 	carv2 "github.com/ipld/go-car/v2"
 	"github.com/ipld/go-car/v2/storage"
 	"github.com/ipld/go-ipld-prime"
@@ -310,7 +310,7 @@ func TestHttpFetch(t *testing.T) {
 				// validate we got the car-scope file form
 				validateCarBody(t, body, srcData.Root, wantCids, false)
 				// validate that we got the full depth form under the path
-				gotDir := unixfs.CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
+				gotDir := CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
 				gotDir.Path = "want0"
 				unixfs.CompareDirEntries(t, srcData.Children[1].Children[1].Children[1], gotDir)
 			}},
@@ -334,7 +334,7 @@ func TestHttpFetch(t *testing.T) {
 				// validate we got the car-scope file form
 				validateCarBody(t, body, srcData.Root, wantCids, false)
 				// validate that we got the full depth form under the path
-				gotDir := unixfs.CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
+				gotDir := CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
 				gotDir.Path = "want0"
 				unixfs.CompareDirEntries(t, srcData.Children[1].Children[1].Children[1], gotDir)
 			}},
@@ -424,7 +424,7 @@ func TestHttpFetch(t *testing.T) {
 				// validate we got the car-scope file form
 				validateCarBody(t, body, srcData.Root, wantCids, false)
 				// validate that we got the full depth form under the path
-				gotDir := unixfs.CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
+				gotDir := CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
 				gotDir.Path = "want0"
 				unixfs.CompareDirEntries(t, srcData.Children[1].Children[1].Children[1], gotDir)
 			}},
@@ -448,7 +448,7 @@ func TestHttpFetch(t *testing.T) {
 				// validate we got the car-scope file form
 				validateCarBody(t, body, srcData.Root, wantCids, false)
 				// validate that we got the full depth form under the path
-				gotDir := unixfs.CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
+				gotDir := CarToDirEntry(t, bytes.NewReader(body), srcData.Children[1].Children[1].Children[1].Root, true)
 				gotDir.Path = "want0"
 				unixfs.CompareDirEntries(t, srcData.Children[1].Children[1].Children[1], gotDir)
 			}},
@@ -741,8 +741,8 @@ func TestHttpFetch(t *testing.T) {
 					if testCase.validateBodies != nil && testCase.validateBodies[i] != nil {
 						testCase.validateBodies[i](t, srcData[i], body)
 					} else {
-						// gotDir := unixfs.CarToDirEntry(t, bytes.NewReader(body), srcData[i].Root, true)
-						gotLsys := unixfs.CarBytesLinkSystem(t, bytes.NewReader(body))
+						// gotDir := CarToDirEntry(t, bytes.NewReader(body), srcData[i].Root, true)
+						gotLsys := CarBytesLinkSystem(t, bytes.NewReader(body))
 						gotDir := unixfs.ToDirEntry(t, gotLsys, srcData[i].Root, true)
 						unixfs.CompareDirEntries(t, srcData[i], gotDir)
 					}
